Small updates, Created NCR Model and Repository.

+++ src/main/java/com/codewranglers/workflowmanager/models/data/JobRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JobRepository extends CrudRepository<Job, Long> {
+public interface JobRepository extends CrudRepository<Job, Integer> {
 }
